file: Move interface from param to file
