+++ file/types.go
@@ -1,4 +1,4 @@
-package param
+package file
 
 import (
 	"context"
